backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.97.1
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.35.0
whitenoise==6.9.0
